requirements.txt
aiogram
python-dotenv
requests
qrcode[pil]
mercadopago
httpx
sqlite3
